requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
telethon==1.34.0
python-dotenv==1.0.0
pydantic==2.5.3
requests==2.31.0
beautifulsoup4==4.12.3
tweepy==4.14.0
yfinance==0.2.36
python-dateutil==2.8.2
supabase==2.9.0
numpy==1.26.3
scikit-learn==1.4.0
langdetect==1.0.9